requirements.txt
django-tables2-bootstrap4
